refactor: change type for Id from int to long

+++ backend/WebAPI/Controllers/SummaryController.cs
@@ -23,7 +23,7 @@ public class SummaryController : BaseController
     }
 
     [HttpGet("{id}"), AllowAnonymous]
-    public async Task<IActionResult> Get([FromRoute] int id)
+    public async Task<IActionResult> Get([FromRoute] long id)
     {
         var query = await _summaryService.GetByIdAsync(id);
         return Ok(query);
@@ -38,14 +38,14 @@ public class SummaryController : BaseController
 
     // PUT api/<controller>/5
     [HttpPut("{id}")]
-    public async Task<IActionResult> Update([FromRoute] int id, [FromBody] SummaryDto summaryDto)
+    public async Task<IActionResult> Update([FromRoute] long id, [FromBody] SummaryDto summaryDto)
     {
         return Ok(await _summaryService.UpdateAsync(id,summaryDto));
     }
     
     // DELETE api/<controller>/5
     [HttpDelete("{id}")]
-    public async Task<IActionResult> Delete(int id)
+    public async Task<IActionResult> Delete(long id)
     {
         await _summaryService.DeleteAsync(id);
         return NoContent();
